requirements.txt
pandas>=2.0.0
pyarrow>=15.0.0
tqdm>=4.65.0
joblib>=1.3.0


# NLP 工具
nltk>=3.9.0
opencc-python-reimplemented>=0.1.7
ckip-transformers>=0.3.4
torch>=2.0.0

rank-bm25>=0.2.2

autogen-agentchat
autogen-ext
ollama==0.4.9